Allow multiple mei versions
In meicommon.h we can now specify a set of versions that libmei will support (currently, 2011-05 and 2012). The default version for new documents is 2012. The mei document version is set based on the meiversion attribute in the mei file.

Also updated the python bindings for MeiDocument constructor. Threw in a unit test as well.