Update Schema Parsing script
This commit removes support for roma-generated builds. Rather, all new schema parsing should be done with a compiled ODD. This was done because a) installing Roma is a pain, and b) it's quite easy to generate a compiled ODD of the latest version at custom.music-encoding.org.

Also, this commit changes the schema parser so that it creates separate sub-directories in the output directory for each language generated.