Fixed: setDocument will throw an exception if the root is not set
Previously, when adding an element to a document it wouldn't matter if the root element was not set.

This fix raises an exception if you try to add an element to a document where it is not set.

The exception, DocumentRootNotSetException, is now thrown when you attempt to add an element to a document with no root. This exception could be caught and the root set before proceeding.