New: Pugixml import supports XML namespaces
Included with a new import test to check that namespaces are correctly imported.

This also fixes a bug where the MeiDocument was not being used to store all the namespaces declared. With this commit, any namespace that is declared in the document is also registered with the MeiDocument. Since namespaces are stored as pointers on attributes and (potentially) elements, this means that there is now a global registry of all namespaces in use on the document.