New: XML Importer
This commit completely re-works the XML Importer. Importing is now done entirely through functions, not a class, and the return value from these functions is an XMLImportResult struct containing the MeiDocument object, the XML Processing Instructions, any Import Warnings, and an import status enum (0 = OK, 1 = not OK)

This new importer can be operated in two modes, "strict" and "lax", set via the MEI_STRICT_IMPORT and MEI_LAX_IMPORT constants. In strict mode the library behaves as before, but in lax mode, the library will import an MEI file even if the elements were not registered via a customization. It will also allow documents without an `meiversion` attribute on the root element.